# cuda : tweak mm stride to double perf on P40 + GTX 970