This simple change more than doubles the prompt processing speed when I am using both of my GPUs. All results are with `-DLLAMA_CUDA_FORCE_MMQ=ON` because I do not have tensor cores.

| GPU      |Model     | Test | t/s master | t/s PR | Speedup |
|----------|-----------|------|------------|--------|---------|
| GTX 970 + P40  |7b q4_0   | pp512   |   217.28 ± 0.004     | 458.18 ± 0.050   |  2.11    |
| GTX 970 + P40  |13b q4_k_s  | pp512   |    111.45 ± 0.001     |  246.14 ± 0.040  |  2.21   |

Can someone with more than one GPU that supports DP4A (compute capability >= 6.1) test with `-DLLAMA_CUDA_FORCE_MMQ=ON` and `-DLLAMA_CUDA_FORCE_MMQ=OFF` (assuming tensor cores are available), in case I need to make this change conditional? Unfortunately, I have only one multi-GPU configuration available to test.

Setting MUL_MAT_SRC1_COL_STRIDE to 1024 is sufficient, if anyone experiences a performance regression with this change, could they test that value too?

cc @Ph0rk0z since you have access to dual P40s.

ref #3814