@ziedbha 2023-11-27
Just curious, do you happen to know how many cuda streams end up being used in either model config? Or maybe an Nsight systems report you'd be willing to share? Would be nice to look at the perf gain from the difference in the stream usage.