@slaren 2023-11-27
Since @JohannesGaessler wrote that code and he uses P40s, I am surprised that the value of `MUL_MAT_SRC1_COL_STRIDE` is not already tuned for the P40. Maybe this is only an improvement with the GTX 970?