@cebtenzzre 2023-11-27
The current version shouldn't have an effect unless you use `-DLLAMA_CUDA_FORCE_MMQ=ON` with a card that doesn't support DP4A. I see no siginficant performance different with just my GTX 970, only when I am using both cards.