@JohannesGaessler 2023-11-27
MMQ needs compute capability 6.1 or higher, otherwise the `__dp4a` instruction is not present. I did not consider the case of one of the GPUs having a compute capability < 6.1 when optimizing the performance via the stride. Such GPUs cannot use MMQ and should fall back to dequantizing the weight matrix and running cuBLAS. Looking at the code it seems that the dequantized weight matrix is not being cached. So for a stride of 128 this would then imply that the weight matrix is being dequantized four times for a batch size of 512 leading to poor performance. That would also explain why for FP16 on Volta or higher the high stride value yields better performance.