@cebtenzzre 2024-03-16
This PR doesn't seem to be relevant anymore - on master with pp512 I'm getting about 654 t/s with my GTX 970 and Tesla P40, with or without this change.